backend/requirements.txt
# === Core Framework ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# === Database ===
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1

# === Data & ML ===
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
xgboost==2.1.3
lightgbm==4.5.0
shap==0.46.0

# === NBA Data ===
nba_api==1.5.2

# === Visualization (EDA) ===
matplotlib==3.10.0
seaborn==0.13.2
jupyter==1.1.1

# === RAG / Intelligence (Phase 4) ===
google-generativeai==0.8.4
chromadb==0.5.23

# === Utilities ===
python-dotenv==1.0.1
pyyaml==6.0.2
httpx==0.28.1
requests==2.32.3

# === Testing ===
pytest==8.3.4
pytest-asyncio==0.25.2
